revision proposal

+++ src/main/java/com/comitfy/crm/app/dto/ProductMaterialDTO.java
@@ -4,11 +4,13 @@ import com.comitfy.crm.util.common.BaseDTO;
 import lombok.Data;
 import net.minidev.json.annotate.JsonIgnore;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductMaterialDTO extends BaseDTO {
 
     @JsonIgnore
     private ProductDTO product;
     private MaterialDTO material;
-    private Integer quantity;
+    private BigDecimal quantity;
 }
